Assist with preparing a Vagrant base box
Getting a CentOS image prepared to be used as a Vagrant base box
actually has a number of gotchas and edge cases.

This script corrects all of these and is designed to be run as root
on the target VM just prior to running vagrant packaging.

Warning: This will disable eth1. If you wish to keep using the VM
create a copy before running this script.